--- src/vs/workbench/contrib/chat/browser/actions/chatContentParser.ts
@@ -1,50 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { marked } from 'vs/base/common/marked/marked';
-
-
-interface MarkdownContent {
-	type: 'markdown';
-	content: string;
-}
-
-interface CodeContent {
-	type: 'code';
-	language: string;
-	content: string;
-}
-
-type Content = MarkdownContent | CodeContent;
-
-export function splitMarkdownAndCodeBlocks(markdown: string): Content[] {
-	const lexer = new marked.Lexer();
-	const tokens = lexer.lex(markdown);
-
-	const splitContent: Content[] = [];
-
-	let markdownPart = '';
-	tokens.forEach((token) => {
-		if (token.type === 'code') {
-			if (markdownPart.trim()) {
-				splitContent.push({ type: 'markdown', content: markdownPart });
-				markdownPart = '';
-			}
-			splitContent.push({
-				type: 'code',
-				language: token.lang || '',
-				content: token.text,
-			});
-		} else {
-			markdownPart += token.raw;
-		}
-	});
-
-	if (markdownPart.trim()) {
-		splitContent.push({ type: 'markdown', content: markdownPart });
-	}
-
-	return splitContent;
-}
